+++ src/jua/weather/_query_engine.py
@@ -11,7 +11,12 @@ from pydantic import validate_call
 from jua._api import QueryEngineAPI
 from jua._utils.remove_none_from_dict import remove_none_from_dict
 from jua.client import JuaClient
-from jua.types.geo import LatLon, PredictionTimeDelta, SpatialSelection
+from jua.types.geo import (
+    LatLon,
+    PredictionTimeDelta,
+    SpatialSelection,
+    validate_unique_point_keys,
+)
 from jua.weather._model_meta import get_model_meta_info
 from jua.weather._stream import process_arrow_streaming_response
 from jua.weather._types.forecast import ForecastData
@@ -328,6 +333,8 @@ class QueryEngine:
         """
         if isinstance(points, LatLon):
             points = [points]
+        if points is not None:
+            validate_unique_point_keys(points)
 
         geo = build_geo_filter(latitude, longitude, points, method)
         model_meta = get_model_meta_info(model)
@@ -382,12 +389,18 @@ class QueryEngine:
         stream: bool = False,
         print_progress: bool | None = None,
     ) -> pd.DataFrame:
-        if payload.geo.type == "point" and payload.geo.method == "bilinear" and stream:
-            logger.warning(
-                "Cannot use streaming responses with bilinear interpolation. Setting "
-                "stream=False."
-            )
-            stream = False
+        group_by = payload.group_by or [
+            "model",
+            "init_time",
+            "prediction_timedelta",
+            "latitude",
+            "longitude",
+        ]
+        if payload.geo.type == "point":
+            group_by.append("point")
+
+        if all(get_model_meta_info(model).has_grid_access for model in payload.models):
+            payload.group_by = group_by
 
         est_requested_points = payload.num_requested_points()
         if est_requested_points > self._MAX_POINTS_PER_REQUEST: