+++ src/jua/weather/models.py
@@ -36,8 +36,6 @@ class Models(str, Enum):
 
     # Without Grid Access
     ECMWF_AIFS_ENSEMBLE = "ecmwf_aifs025_ensemble"
-    ECMWF_AIFS_SINGLE = "ecmwf_aifs025_single"
-    ECMWF_IFS_ENSEMBLE = "ecmwf_ifs025_ensemble"
     GFS_GLOBAL_ENSEMBLE = "gfs_global_ensemble"
     GFS_GLOBAL_SINGLE = "gfs_global_single"
     GFS_GRAPHCAST = "gfs_graphcast025"